+++ src/signup/index.ts
@@ -1,0 +1,4 @@
+import Signup from "./signup";
+import SignupRouter from "./signup.router";
+
+export { Signup, SignupRouter };
