+++ src/signup/signup.ts
@@ -1,0 +1,9 @@
+import Vue from "vue";
+import Component from "vue-class-component";
+
+@Component
+export default class Signup extends Vue {
+  private signup() {
+    console.log("signup");
+  }
+}
